add comment

+++ app/src/main/java/com/example/java_squad/Count.java
@@ -1,29 +1,49 @@
 package com.example.java_squad;
 
 import java.util.Date;
-
+/**
+ * Count class.
+ */
 public class Count extends Trial{
     private String object;
     private Integer count;
-
-    public Count(String experimenter, Date experiment_date, String object, double count) {
+    /**
+     * Constructor for Count class.
+     * @param experimenter
+     * Person who add the new trial to the experiment.
+     * @param experiment_date
+     * Date to create the new trial
+     * @param object
+     * The object for experiment
+     * @param count
+     * The experiment result
+     */
+    public Count(String experimenter, Date experiment_date, String object, Integer count) {
         super(experimenter, experiment_date);
         this.object = object;
         this.count = count;
     }
-
+    /**
+     * get experiment object
+     */
     public String getObject() {
         return object;
     }
-
+    /**
+     * set experiment object
+     */
     public void setObject(String object) {
         this.object = object;
     }
-
+    /**
+     * get experiment count
+     */
     public Integer getCount() {
         return count;
     }
-
+    /**
+     * set experiment count
+     */
     public void setCount(Integer count) {
         this.count = count;
     }
